+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.3.1
 
+- Update attribute config defaults
 - Update Fabric Loader to 15+ for embedded MixinExtras
 - Add French translation, thanks to valentin56610